Convert output to uint8.

+++ src/koopaflows/segmentation/other_threshold_segmentation_flow.py
@@ -38,7 +38,7 @@ def segment_other_task(
         )
 
     result.set_data(
-        mask
+        mask.astype(np.uint8)
     )
 
     return result
